Update Refined Storage for 1.12 changes. (#22)
* Updated Refined Storage textures for 1.12 Changes

Removed Processing Pattern Encoder Textures.  That block was removed in Refined Storage 1.5.6.  The earliest 1.12 version of Refined Storage on CurseForge is 1.5.10, and the earliest on 1.12.2 is 1.5.18.
 Functionality was moved to the Pattern Grid.

Updated "pattern_grid.png" to the new layout.  Added the new "pattern_grid_processing.png" in the same style.

* Added RS CTM Glow

Adds textures and files required to satisfy Refined Storage's Connected Textures Mod integration.

Refined Storage added CTM integration in 1.5.14.  Without these new textures and MCMeta files, portions of the default 16x16 textures are applied over top of the 32x32 textures when CTM is installed.

CTM is required by Chisel and a handful of other mods, so it is installed quite often.